Restructure some files and module names
